Fix to_parquet tests on HDFS
