+++ .github/workflows/platform-ci.yml
@@ -110,8 +110,7 @@ jobs:
 
           if [ -n "$WITH_HADOOP" ]; then
             source $CONDA/bin/activate test
-            pytest $PYTEST_CONFIG --cov-config .coveragerc-threaded mars/dataframe/datasource/tests/test_hdfs.py \
-            mars/deploy/yarn
+            pytest $PYTEST_CONFIG --cov-config .coveragerc-threaded -m hadoop
             coverage report
           fi
           if [ -n "$WITH_KUBERNETES" ]; then
