+++ mars/deploy/yarn/tests/test_yarn.py
@@ -28,14 +28,14 @@ import numpy as np
 from numpy.testing import assert_array_equal
 
 import mars.tensor as mt
-from mars.tests.core import flaky
+from mars.tests.core import flaky, require_hadoop
 from mars.deploy.yarn import new_cluster
 
 logger = logging.getLogger(__name__)
 MARS_ROOT = os.path.dirname(os.path.dirname(os.path.dirname(mt.__file__)))
 
 
-@unittest.skipIf(not os.environ.get('HADOOP_HOME'), 'Only run when hadoop is installed')
+@require_hadoop
 @flaky(max_runs=3)
 class Test(unittest.TestCase):
     def tearDown(self):
